requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.5
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
exceptiongroup==1.1.0
google-play-scraper==1.2.2
google-search-results==2.4.1
h11==0.14.0
idna==3.4
lxml==4.9.2
mongoengine==0.25.0
outcome==1.2.0
packaging==23.0
pycparser==2.21
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==0.21.0
python-snappy==0.6.1
pytz==2022.7
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
webdriver-manager==3.8.5
winkerberos==0.9.1
wsproto==1.2.0
